Progress on wdl

+++ db.py
@@ -2,6 +2,7 @@ import psycopg2
 import psycopg2.extras
 import os
 import json
+import models
 
 def connectToDb():
     conn = psycopg2.connect("host=rough-leaf-4031-db.flycast port=5432 dbname=vocabdb user=vocabuser password=vocabpass")
@@ -20,6 +21,20 @@ def get_word(word):
     word_data = record["data"]
     return word_data
 
+def get_saved_words_list_id_by_user_id(user_id):
+    conn = connectToDb()
+    cur = conn.cursor(cursor_factory=psycopg2.extras.DictCursor)
+    cur.execute("select * from word_defn_lists where user_id = %s and name = 'My saved words'", [user_id])
+    d = cur.fetchone()
+    return d[0]
+
+def get_word_defn_id_by_word(word):
+    conn = connectToDb()
+    cur = conn.cursor(cursor_factory=psycopg2.extras.DictCursor)
+    cur.execute("select * from word_defns where word=%s", [word])
+    d = cur.fetchone()
+    return d[0]
+
 def save_word(word, word_data):
     conn = connectToDb()
     word_data = json.dumps(word_data) #convert from dict to str for saving
@@ -32,11 +47,19 @@ def save_word_for_user(user_id, word):
     cur.execute("select * from saved_words where user_id = %s and word = %s", [user_id, word])
     if not cur.fetchone():
         cur.execute("insert into saved_words (user_id, word) values (%s, %s)", [user_id, word])
+    saved_words_list_id = get_saved_words_list_id_by_user_id(user_id)
+    word_defn_id = get_word_defn_id_by_word(word)
+    cur.execute("insert into word_defn_list_map (word_defn_list_id, word_defn_id) values (%s, %s)", [saved_words_list_id, word_defn_id])
 
 def unsave_word_for_user(user_id, word):
     conn = connectToDb()
     cur = conn.cursor(cursor_factory=psycopg2.extras.DictCursor)
-    cur.execute("delete from saved_words where user_id = %s and word = %s", [user_id, word])
+    cur.execute("select * from saved_words where user_id = %s and word = %s", [user_id, word])
+    if cur.fetchone():
+        cur.execute("delete from saved_words where user_id = %s and word = %s", [user_id, word])
+    saved_words_list_id = get_saved_words_list_id_by_user_id(user_id)
+    word_defn_id = get_word_defn_id_by_word(word)
+    cur.execute("delete from word_defn_list_map where word_defn_list_id = %s and word_defn_id = %s", [saved_words_list_id, word_defn_id])
 
 def get_saved_words(user_id):
     conn = connectToDb()
@@ -46,7 +69,7 @@ def get_saved_words(user_id):
     users_saved_words = users_saved_words
     return users_saved_words
 
-def word_saved_by_user_or_not(user_id, word):
+def is_word_saved_by_user(user_id, word):
     conn = connectToDb()
     cur = conn.cursor(cursor_factory=psycopg2.extras.DictCursor)
     cur.execute("select word from saved_words where user_id = %s and word = %s", [user_id, word])
